:sparkles: Support Django Channels layer messages
Channel layers, as implemented in Django Channels, depend on having
control of the program flow. When messages arrive they are handed off
to a dispatch method on the consumer. With Strawberry subscriptions,
the control flow ends up in the subscription resolver. The consequence
of this is that no messages will ever end up consumed by the channel.

This work provides a simple helper to consume channel layer messages
from within the Strawberry subscription resolver. I've also added a
helper to send a message to a group.